examples/daily-market-dashboard/requirements.txt
# App framework
streamlit>=1.42.0
claude-agent-sdk>=0.1.35
python-dotenv>=1.0.0

# Shared by most skills (FMP API calls, data fetching)
requests>=2.31.0

# Theme Detector dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
finvizfinance>=0.14.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
